content/html/document/test/test_bug512367.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=512367
-->
<head>
  <title>Test for Bug 512367</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=512367">Mozilla Bug 512367</a>
<p id="display">
  <iframe src="bug369370-popup.png" id="i" style="width:200px; height:200px"></iframe>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

var frame = document.getElementById("i");

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
  var viewer =
    frame.contentWindow
         .QueryInterface(Components.interfaces.nsIInterfaceRequestor)
         .getInterface(Components.interfaces.nsIWebNavigation)
         .QueryInterface(Components.interfaces.nsIDocShell)
         .contentViewer
         .QueryInterface(Components.interfaces.nsIMarkupDocumentViewer);

  viewer.fullZoom = 1.5;

  setTimeout(function() {
    synthesizeMouse(frame, 30, 30, {});

    netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');
    is(viewer.fullZoom, 1.5, "Zoom in the image frame should not have been reset");

    SimpleTest.finish();
  }, 0);
});
</script>
</pre>
</body>
</html>
